modification routes post, put, delete

+++ backend/controllers/sauce.js
@@ -16,7 +16,7 @@ exports.createSauce = (req, res, next) => {
     userLiked: [],
     userDisliked: [],
   });
- 
+
   sauce
     .save()
     .then(() => res.status(201).json({ message: "Sauce enregistrée !" }))
@@ -44,25 +44,42 @@ exports.modifySauce = (req, res, next) => {
         }`,
       }
     : { ...req.body };
-  Sauce.updateOne(
-    { _id: req.params.id },
-    { ...sauceObject, _id: req.params.id }
-  )
-    .then(() => res.status(200).json({ message: "Elément modifié" }))
-    .catch((error) => res.status(400).json(error));
+
+  delete sauceObject.userId;
+  Sauce.findOne({ _id: req.params.id })
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({ message: "Non-autorisé" });
+      } else {
+        const filename = sauce.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Sauce.updateOne(
+            { _id: req.params.id },
+            { ...sauceObject, _id: req.params.id }
+          )
+            .then(() => res.status(200).json({ message: "Elément modifié" }))
+            .catch((error) => res.status(401).json({ error }));
+        });
+      }
+    })
+    .catch((error) => res.status(400).json({ error }));
 };
 
 exports.deleteSauce = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id })
     .then((sauce) => {
-      const filename = sauce.imageUrl.split("/images/")[1];
-      fs.unlink(`images/${filename}`, () => {
-        Sauce.deleteOne({ _id: req.params.id })
-          .then(() => res.status(200).json({ message: "Elément supprimé" }))
-          .catch((error) => res.status(404).json({ error }));
-      });
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({ message: "Non-autorisé" });
+      } else {
+        const filename = sauce.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Sauce.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: "Elément supprimé" }))
+            .catch((error) => res.status(404).json({ error }));
+        });
+      }
     })
-    .catch((error) => res.status(500).json(error));
+    .catch((error) => res.status(500).json({ error }));
 };
 
 exports.postLikes = (req, res, next) => {
@@ -98,7 +115,7 @@ exports.postLikes = (req, res, next) => {
       sauce
         .save()
         .then(() => res.status(201).json({ message: "Like" }))
-        .catch((error) => res.status(400).json(error));
+        .catch((error) => res.status(400).json({ error }));
     })
-    .catch((error) => res.status(500).json(error));
+    .catch((error) => res.status(500).json({ error }));
 };
